pyproject.toml
[tool.poetry]
name = "analytics-bench"
version = "0.1.0"
description = "LLM Analytics Benchmark for testing how well LLMs handle structured data queries"
authors = ["Your Name <your.email@example.com>"]
readme = "readme.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.3"
tiktoken = "^0.8.0"
kaggle = "^1.6.17"
tqdm = "^4.67.1"
python-dotenv = "^1.0.1"
openai = "^1.59.3"
pydantic = "^2.10.4"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
black = "^24.10.0"
isort = "^5.13.2"
mypy = "^1.14.1"
pandas-stubs = "^2.2.3.241126"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

